pyproject.toml
[tool.poetry]
name = "splitgraph"
version = "0.1.0-pre"
description = "Command line library and Python client for Splitgraph, a version control system for data"
license = "Apache 2.0 modified with Commons Clause"
authors = ["Splitgraph Limited"]
readme = "README.md"
homepage = "https://www.splitgraph.com"
repository = "https://github.com/splitgraph/splitgraph"

[tool.poetry.dependencies]
python = ">=3.6"
click = "^7"
psycopg2-binary = "^2"
parsimonious = "^0.8"
minio = ">=4"
pglast = ">=1.6"
requests = ">=2.22"
docker = ">=4.0"

# Extra requirements for ingestion (CSV)
pandas = {version = ">=0.24", extras = ["ingestion"], optional = true }
sqlalchemy = { version = "^1.3", extras = ["ingestion"], optional = true }
click_log = "^0.3.2"
tqdm = "^4.36.0"


[tool.poetry.dev-dependencies]
pytest = ">=4.4"
pyfakefs = "^3.6"
pytest-cov = "^2.8"
pytest-env = "^0.6"
coveralls = "^1.7"
pyinstaller = "^3.4"
sphinx = "^2.0"
sphinx_rtd_theme = "^0.4.2"
pylint = "^2.4"
black = { version = "*", allows-prereleases = true }
pre-commit = "^1.18"
httpretty = "^0.9.6"
mypy = ">=0.720"

[tool.poetry.extras]
ingestion = ["pandas", "sqlalchemy"]

[tool.poetry.scripts]
sgr = "splitgraph.commandline:cli"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

